backend/requirements.txt
langchain
langgraph
langchain-google-genai
google-genai
pydantic
python-dotenv
fastapi
uvicorn
python-multipart
pypdf2
docx
supabase